wip refactor login
