+++ front/src/modules/ui/layout/components/DefaultLayout.tsx
@@ -1,11 +1,18 @@
+import { useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
 import styled from '@emotion/styled';
-import { useRecoilState, useRecoilValue } from 'recoil';
+import { AnimatePresence, LayoutGroup } from 'framer-motion';
+import { useRecoilValue } from 'recoil';
 
-import { currentUserState } from '@/auth/states/currentUserState';
+import { AuthModal } from '@/auth/components/ui/Modal';
+import { useOnboardingStatus } from '@/auth/hooks/useOnboardingStatus';
+import { OnboardingStatus } from '@/auth/utils/getOnboardingStatus';
 import { CommandMenu } from '@/command-menu/components/CommandMenu';
 import { NavbarAnimatedContainer } from '@/ui/navbar/components/NavbarAnimatedContainer';
 import { MOBILE_VIEWPORT } from '@/ui/themes/themes';
 import { AppNavbar } from '~/AppNavbar';
+import { useIsMatchingLocation } from '~/hooks/useIsMatchingLocation';
+import { CompaniesMockMode } from '~/pages/companies/CompaniesMockMode';
 
 import { isNavbarOpenedState } from '../states/isNavbarOpenedState';
 
@@ -38,22 +45,64 @@ type OwnProps = {
 };
 
 export function DefaultLayout({ children }: OwnProps) {
-  const currentUser = useRecoilState(currentUserState);
-  const userIsAuthenticated = !!currentUser;
+  const navigate = useNavigate();
+  const isMatchingLocation = useIsMatchingLocation();
+
+  const onboardingStatus = useOnboardingStatus();
+  useEffect(() => {
+    if (onboardingStatus === OnboardingStatus.OngoingUserCreation) {
+      if (
+        !isMatchingLocation('/sign-in') &&
+        !isMatchingLocation('/sign-up') &&
+        !isMatchingLocation('/invite') &&
+        !isMatchingLocation('/verify')
+      ) {
+        navigate('/sign-in');
+      }
+    } else if (onboardingStatus === OnboardingStatus.OngoingWorkspaceCreation) {
+      if (!isMatchingLocation('/create/workspace')) {
+        navigate('/create/workspace');
+      }
+    } else if (onboardingStatus === OnboardingStatus.OngoingProfileCreation) {
+      if (!isMatchingLocation('/create/profile')) {
+        navigate('/create/profile');
+      }
+    } else if (onboardingStatus === OnboardingStatus.Completed) {
+      if (
+        isMatchingLocation('/sign-in') ||
+        isMatchingLocation('/sign-up') ||
+        isMatchingLocation('/invite') ||
+        isMatchingLocation('/verify') ||
+        isMatchingLocation('/create/workspace') ||
+        isMatchingLocation('/create/profile')
+      ) {
+        navigate('/');
+      }
+    }
+  }, [onboardingStatus, navigate, isMatchingLocation]);
 
   return (
     <StyledLayout>
-      {userIsAuthenticated ? (
-        <>
-          <CommandMenu />
-          <NavbarAnimatedContainer>
-            <AppNavbar />
-          </NavbarAnimatedContainer>
-          <MainContainer>{children}</MainContainer>
-        </>
-      ) : (
-        children
-      )}
+      <>
+        <CommandMenu />
+        <NavbarAnimatedContainer>
+          <AppNavbar />
+        </NavbarAnimatedContainer>
+        <MainContainer>
+          {onboardingStatus !== OnboardingStatus.Completed ? (
+            <>
+              <CompaniesMockMode />
+              <AnimatePresence mode="wait">
+                <LayoutGroup>
+                  <AuthModal>{children}</AuthModal>
+                </LayoutGroup>
+              </AnimatePresence>
+            </>
+          ) : (
+            <>{children}</>
+          )}
+        </MainContainer>
+      </>
     </StyledLayout>
   );
 }
