+++ front/src/pages/auth/__stories__/Index.stories.tsx
@@ -5,25 +5,25 @@ import { AuthLayout } from '@/ui/layout/components/AuthLayout';
 import { graphqlMocks } from '~/testing/graphqlMocks';
 import { getRenderWrapperForPage } from '~/testing/renderWrappers';
 
-import { Index } from '../Index';
+import { SignInUp } from '../SignInUp';
 
-const meta: Meta<typeof Index> = {
-  title: 'Pages/Auth/Index',
-  component: Index,
+const meta: Meta<typeof SignInUp> = {
+  title: 'Pages/Auth/SignInUp',
+  component: SignInUp,
 };
 
 export default meta;
 
-export type Story = StoryObj<typeof Index>;
+export type Story = StoryObj<typeof SignInUp>;
 
 export const Default: Story = {
   render: getRenderWrapperForPage(
     <AuthLayout>
       <AuthModal>
-        <Index />
+        <SignInUp />
       </AuthModal>
     </AuthLayout>,
-    '/auth',
+    '/',
   ),
   parameters: {
     msw: graphqlMocks,
