+++ server/.env.example
@@ -6,6 +6,6 @@ LOGIN_TOKEN_EXPIRES_IN=15m
 REFRESH_TOKEN_SECRET=secret_refresh_token
 REFRESH_TOKEN_EXPIRES_IN=90d
 PG_DATABASE_URL=postgres://postgres:postgrespassword@localhost:5432/default?connection_limit=1
-FRONT_AUTH_CALLBACK_URL=http://localhost:3001/auth/callback
+FRONT_AUTH_CALLBACK_URL=http://localhost:3001/verify
 STORAGE_TYPE=local
 STORAGE_LOCAL_PATH=.local-storage
